Extract out config types
