+++ packages/core/src/lib/config.types.mts
@@ -1,0 +1,9 @@
+export type AsyncPromptConfig = {
+  message: string | Promise<string> | (() => Promise<string>);
+  validate?: (value: string) => boolean | string | Promise<string | boolean>;
+};
+
+export type ResolvedPromptConfig = {
+  message: string;
+  validate: (value: string) => boolean | string | Promise<string | boolean>;
+};
